+++ include/Interpret.cuh
@@ -1,5 +1,5 @@
-__host__ double TellNbOrbits (double time);
-__host__ double TellNbOutputs (double time);
+__host__ float TellNbOrbits (float time);
+__host__ float TellNbOutputs (float time);
 __host__ void SearchVariable (char *name, char *stringval);
 __host__ void ReadVariables (char *filename);
 __host__ void TellEverything ();
