+++ include/Viscosity.cuh
@@ -1,6 +1,6 @@
 __host__ void InitViscosityDevice ();
 __host__ void InitViscosity ();
-__host__ void UpdateVelocitiesWithViscosity(double *Vrad, double *Vtheta, double *Dens, double DeltaT);
-__host__ void ComputeViscousTerms (double *Vrad, double *Vtheta, double *Dens);
-__host__ double FViscosity (double r);
-__host__ double AspectRatioHost(double r);
+__host__ void UpdateVelocitiesWithViscosity(float *Vrad, float *Vtheta, float *Dens, float DeltaT);
+__host__ void ComputeViscousTerms (float *Vrad, float *Vtheta, float *Dens);
+__host__ float FViscosity (float r);
+__host__ float AspectRatioHost(float r);
