+++ src/SideEuler.cu
@@ -3,18 +3,18 @@
 extern int OpenInner, NSEC, size_grid, NonReflecting, Adiabatic, NRAD;
 extern int Evanescent, SelfGravity, ExcludeHill, dimfxy, OuterSourceMass;
 
-extern double *SigmaMed, *SoundSpeed, *EnergyMed, *mean_dens, *mean_energy;
-extern double *cs0, *cs1, *csnrm1, *csnrm2, *mean_dens2, *mean_energy2, *viscosity_array;
+extern float *SigmaMed, *SoundSpeed, *EnergyMed, *mean_dens, *mean_energy;
+extern float *cs0, *cs1, *csnrm1, *csnrm2, *mean_dens2, *mean_energy2, *viscosity_array;
 
-extern double ADIABATICINDEX, FLARINGINDEX, SIGMASLOPE, ASPECTRATIO;
-extern double TRANSITIONWIDTH, TRANSITIONRATIO, TRANSITIONRADIUS, PhysicalTime, PhysicalTimeInitial;
-extern double LAMBDADOUBLING;
+extern float ADIABATICINDEX, FLARINGINDEX, SIGMASLOPE, ASPECTRATIO;
+extern float TRANSITIONWIDTH, TRANSITIONRATIO, TRANSITIONRADIUS, PhysicalTime, PhysicalTimeInitial;
+extern float LAMBDADOUBLING;
 
-extern double *Vrad_d, *Dens_d, *Energy_d, *SoundSpeed_d, *mean_dens_d, *mean_energy_d;
-extern double *cs0_d, *cs1_d, *csnrm1_d, *csnrm2_d, *mean_dens_d2, *mean_energy_d2, *viscosity_array_d;
-extern double *Vtheta_d, *SigmaMed_d, *EnergyMed_d, *GLOBAL_bufarray_d, *VthetaInt_d, *Work_d, *DensInt_d, *Qbase_d;
+extern float *Vrad_d, *Dens_d, *Energy_d, *SoundSpeed_d, *mean_dens_d, *mean_energy_d;
+extern float *cs0_d, *cs1_d, *csnrm1_d, *csnrm2_d, *mean_dens_d2, *mean_energy_d2, *viscosity_array_d;
+extern float *Vtheta_d, *SigmaMed_d, *EnergyMed_d, *GLOBAL_bufarray_d, *VthetaInt_d, *Work_d, *DensInt_d, *Qbase_d;
 
-double  mean_dens_r, mean_energy_r, mean_dens_r2, mean_energy_r2, cs0_r, cs1_r, csnrm1_r, *CellAbscissa,        \
+float  mean_dens_r, mean_energy_r, mean_dens_r2, mean_energy_r2, cs0_r, cs1_r, csnrm1_r, *CellAbscissa,        \
 csnrm2_r, *CellAbscissa_d, *CellOrdinate, *CellOrdinate_d, *Vmoy_d;
 
 extern dim3 dimGrid, dimBlock, dimBlock2, dimGrid2;
@@ -22,7 +22,7 @@ extern dim3 dimGrid, dimBlock, dimBlock2, dimGrid2;
 extern double OmegaFrame, *Rinf, *Rmed, *Rmed_d;
 
 
-__host__ void ApplyBoundaryCondition (double *Dens, double *Energy, double *Vrad, double *Vtheta, double step)
+__host__ void ApplyBoundaryCondition (float *Dens, float *Energy, float *Vrad, float *Vtheta, float step)
 {
   if(OpenInner == YES) OpenBoundary ();
 
@@ -36,22 +36,22 @@ __host__ void ApplyBoundaryCondition (double *Dens, double *Energy, double *Vrad
 
 
 
-__host__ void NonReflectingBoundary (double *Dens, double *Energy, double *Vrad)
+__host__ void NonReflectingBoundary (float *Dens, float *Energy, float *Vrad)
 {
   int i,i_angle, i_angle2;
-  double dangle, dangle2;
+  float dangle, dangle2;
 
   ReduceCs();
 
   i = 1;
   dangle = (pow(Rinf[i],-1.5)-1.0)/(.5*(cs0_r+cs1_r));
   dangle *= (Rmed[i] - Rmed[i-1]);
-  i_angle = (int)(dangle/2.0/PI*(double)NSEC+.5);
+  i_angle = (int)(dangle/2.0/PI*(float)NSEC+.5);
 
   i = NRAD-1;
   dangle2 = (pow(Rinf[i-1],-1.5)-1.0)/(.5*(csnrm1_r+csnrm2_r));
   dangle2 *= (Rmed[i]-Rmed[i-1]);
-  i_angle2 = (int)(dangle2/2.0/PI*(double)NSEC+.5);
+  i_angle2 = (int)(dangle2/2.0/PI*(float)NSEC+.5);
 
   NonReflectingBoundaryKernel<<<dimGrid, dimBlock>>>(Dens_d, Energy_d, i_angle, NSEC, Vrad_d, SoundSpeed_d, SigmaMed[1], NRAD,
   SigmaMed[NRAD-2], i_angle2);
@@ -78,7 +78,7 @@ __host__ void ReduceCs ()
 }
 
 
-__host__ void ReduceMean (double *Dens, double *Energy)
+__host__ void ReduceMean (float *Dens, float *Energy)
 {
   ReduceMeanKernel<<<dimGrid, dimBlock>>>(Dens_d, Energy_d, NSEC, mean_dens_d, mean_energy_d, mean_dens_d2, mean_energy_d2, NRAD);
   gpuErrchk(cudaDeviceSynchronize());
@@ -91,9 +91,9 @@ __host__ void ReduceMean (double *Dens, double *Energy)
 
 
 
-__host__ void EvanescentBoundary (double *Vrad, double *Vtheta, double *Dens, double *Energy, double step)
+__host__ void EvanescentBoundary (float *Vrad, float *Vtheta, float *Dens, float *Energy, float step)
 {
-  double Tin, Tout, DRMIN, DRMAX;
+  float Tin, Tout, DRMIN, DRMAX;
   /* Orbital period at inner and outer boundary */
   Tin = 2.0*PI*pow(Rmed[0],3./2);;
   Tout = 2.0*PI*pow(Rmed[NRAD-1],3./2);
@@ -102,7 +102,7 @@ __host__ void EvanescentBoundary (double *Vrad, double *Vtheta, double *Dens, do
   DRMAX = Rmed[NRAD-1]*0.84;
 
   for (int i = 0; i < NRAD; i++) viscosity_array[i] = FViscosity(Rmed[i]);
-  gpuErrchk(cudaMemcpy(viscosity_array_d,viscosity_array, (NRAD+1)*sizeof(double), cudaMemcpyHostToDevice));
+  gpuErrchk(cudaMemcpy(viscosity_array_d,viscosity_array, (NRAD+1)*sizeof(float), cudaMemcpyHostToDevice));
 
   EvanescentBoundaryKernel<<<dimGrid2, dimBlock2>>>(Rmed_d, Vrad_d, Vtheta_d, Energy_d, Dens_d,
     viscosity_array_d, DRMIN, DRMAX, NRAD, NSEC, Tin, Tout, step, SIGMASLOPE, FLARINGINDEX,  GLOBAL_bufarray_d,
@@ -121,7 +121,7 @@ __host__ void OpenBoundary ()
 
 
 
-__host__ Pair ComputeAccel (Force *force, double *Dens, double x, double y, double rsmoothing, double mass)
+__host__ Pair ComputeAccel (Force *force, float *Dens, float x, float y, float rsmoothing, float mass)
 {
   Pair acceleration;
   ComputeForce (force, Dens, x, y, rsmoothing, mass, dimfxy, 0.0, 0.0);
@@ -138,7 +138,7 @@ __host__ Pair ComputeAccel (Force *force, double *Dens, double x, double y, doub
 
 
 
-__host__ void DivisePolarGrid (double *Qbase_d, double *DensInt_d, double *Work_d)
+__host__ void DivisePolarGrid (float *Qbase_d, float *DensInt_d, float *Work_d)
 {
   DivisePolarGridKernel<<<dimGrid2, dimBlock2>>> (Qbase_d, DensInt_d, Work_d, NRAD, NSEC);
   gpuErrchk(cudaDeviceSynchronize());
@@ -148,8 +148,8 @@ __host__ void DivisePolarGrid (double *Qbase_d, double *DensInt_d, double *Work_
 
 __host__ void InitComputeAccel ()
 {
-  CellAbscissa    = (double *)malloc(size_grid*sizeof(double));
-  CellOrdinate    = (double *)malloc(size_grid*sizeof(double));
+  CellAbscissa    = (float *)malloc(size_grid*sizeof(float));
+  CellOrdinate    = (float *)malloc(size_grid*sizeof(float));
 
   InitComputeAccelDevice();
 
@@ -161,15 +161,15 @@ __host__ void InitComputeAccel ()
 
 __host__ void InitComputeAccelDevice()
 {
-  gpuErrchk(cudaMalloc((void**)&CellAbscissa_d, size_grid*sizeof(double)));
-  gpuErrchk(cudaMalloc((void**)&CellOrdinate_d, size_grid*sizeof(double)));
-  //gpuErrchk(cudaMemcpy(CellAbscissa_d, CellAbscissa, size_grid*sizeof(double), cudaMemcpyHostToDevice));
-  //gpuErrchk(cudaMemcpy(CellOrdinate_d, CellOrdinate, size_grid*sizeof(double), cudaMemcpyHostToDevice));
+  gpuErrchk(cudaMalloc((void**)&CellAbscissa_d, size_grid*sizeof(float)));
+  gpuErrchk(cudaMalloc((void**)&CellOrdinate_d, size_grid*sizeof(float)));
+  //gpuErrchk(cudaMemcpy(CellAbscissa_d, CellAbscissa, size_grid*sizeof(float), cudaMemcpyHostToDevice));
+  //gpuErrchk(cudaMemcpy(CellOrdinate_d, CellOrdinate, size_grid*sizeof(float), cudaMemcpyHostToDevice));
 }
 
 
 
-__host__ void CorrectVtheta (double *Vtheta, double domega)
+__host__ void CorrectVtheta (float *Vtheta, float domega)
 {
   CorrectVthetaKernel<<<dimGrid2, dimBlock2>>>(Vtheta_d, domega, Rmed_d, NRAD, NSEC);
   gpuErrchk(cudaDeviceSynchronize());
@@ -177,9 +177,9 @@ __host__ void CorrectVtheta (double *Vtheta, double domega)
 
 
 
-__host__ void ApplySubKeplerianBoundary(double *VthetaInt)
+__host__ void ApplySubKeplerianBoundary(float *VthetaInt)
 {
-  double VKepIn, VKepOut;
+  float VKepIn, VKepOut;
 
   if (!SelfGravity){
     VKepIn = sqrt(G*1.0/Rmed[0] * (1.0 - (1.0+SIGMASLOPE-2.0*FLARINGINDEX) * \
